refactor(scaling): extract pure scaling helpers into linopy/scaling.py
Move validate_scaling, ensure_scaling, variable_solver_scaling and the
variable/constraint scaling lookups out of common.py into a dedicated
module and repoint all import sites.
